requirements.txt
fastapi
uvicorn
psycopg
chromadb
sentence-transformers
google-generativeai
python-dotenv
langgraph